pyproject.toml
[project]
name = "mdzen"
version = "0.2.0"
description = "MDZen - AI Agent for Molecular Dynamics Setup (MD + 膳/禅)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # MCP Server (FastMCP)
    "fastmcp>=0.1.0",
    "mcp[cli]>=1.18.0",

    # Google Agent Development Kit (ADK)
    "google-adk>=0.2.0",
    "litellm>=1.0.0",

    # Structure prediction & modeling
    # Note: Install openmm, rdkit, mdanalysis, biopython, pandas, numpy, scipy, pdbfixer via conda first
    # conda install -c conda-forge openmm rdkit mdanalysis biopython pandas numpy scipy openblas pdbfixer
    # Boltz-2: Install manually when needed (Phase 2-3)
    #   pip install 'boltz[cuda]' --no-deps
    #   Then install missing dependencies individually if needed
    "pdb2pqr>=3.1.0",
    "propka>=3.5.0",
    "gemmi>=0.6.0",  # mmCIF parsing for Boltz-2 output
    "dimorphite-dl>=1.3.2",  # pH-dependent protonation state generation

    # Topology manipulation
    "parmed>=4.3.0",

    # LLM integration
    "anthropic>=0.40.0",
    "prompt-toolkit>=3.0.0",

    # Data models & validation
    "pydantic>=2.12.3",

    # CLI & UI
    "typer>=0.19.2",
    "rich>=14.2.0",

    # Utilities
    "httpx>=0.28.1",
    "requests>=2.31.0",
    "pyyaml>=6.0",

    # Notebook development
    "jupyter>=1.0.0",
    "ipykernel>=6.29.0",
    "python-dotenv>=1.0.0",

    # Visualization & analysis (for test notebooks)
    "py3Dmol>=2.0.0",
    "mdtraj>=1.9.0",
]

[project.optional-dependencies]
# Heavy dependencies (install via conda)
heavy = [
    "openmm>=8.3.1",
    "mdanalysis>=2.10.0",
    "rdkit>=2023.9.1",
    "biopython>=1.83",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[project.scripts]
mdzen = "main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["servers", "common", "src/mdzen"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
